# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Church Records Backend REST API
host: localhost:10120
...
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /church:
    # binds a127 app logic to a route
    x-swagger-router-controller: church.controller
    get:
      tags: [Church]
      description: Returns list of churchs from Church Collection
      # used as the method name of the controller
      operationId: getAllChurch
      responses:
        "200":
          description: Church Found
          schema:
            type: array
            items:
              $ref: "#/definitions/church"
        "404":
          description: Empty Collection
          schema:
            $ref: "#/definitions/message"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    post:
      tags: [Church]
      description: Creates new church entry
      consumes:
        - application/json
      produces:
        - application/json
      operationId: createChurch
      parameters:
        - name: churchdata
          required: true
          in: body
          description: a new church details
          schema:
            $ref: "#/definitions/church"
      responses:
        "200":
          description: New church Created
          schema:
            $ref: "#/definitions/church"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Church]
      description: Remove all churches in database
      operationId: removeAllChurch
      responses:
        "200":
          description: Successfully removed all churches
          schema:
            $ref: "#/definitions/message"
        default:
          description: Error
          schema:
            $ref: "#/definitions/message"
  /church/{churchid}:
    x-swagger-router-controller: church.controller
    get:
      tags: [Church]
      description: Return single church by given id
      operationId: getSingleChurch
      parameters:
        - name: churchid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Church Found
          schema:
            $ref: "#/definitions/church"
        "404":
          description: Church not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Church]
      description: Remove single church by given id
      operationId: removeSingleChurch
      parameters:
        - name: churchid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Successfully removed
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    patch:
      tags: [Church]
      description: Update by church id
      operationId: updateSingleChurch
      parameters:
        - name: churchid
          in: path
          required: true
          type: string
        - name: church
          in: body
          description: The Updated church info
          required: true
          schema:
            $ref: "#/definitions/church"
      responses:
        "200":
          description: Church Found and Updated
          schema:
            $ref: "#/definitions/church"
        "404":
          description: Church not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
#Path to retrieve All communities of specific Church
  /church/{churchid}/community:
    x-swagger-router-controller: church.controller
    get:
      tags: [Church]
      description: Return All Communities by given churchcode
      operationId: getChurchCommunity
      parameters:
        - name: churchid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Community Found
          schema:
            type: array
            items:
              $ref: "#/definitions/community"
        "404":
          description: Community not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"

#Path to retrieve All contributions of specific Church
  /church/{churchid}/contribution:
    x-swagger-router-controller: church.controller
    get:
      tags: [Church]
      description: Return All Contributions by given churchcode
      operationId: getChurchContribution
      parameters:
        - name: churchid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Contribution Found
          schema:
            type: array
            items:
              $ref: "#/definitions/contribution"
        "404":
          description: Contribution not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
#Path to retrieve church Believers
  /church/{churchid}/believers:
    x-swagger-router-controller: church.controller
    get:
      tags: [Church]
      description: Return single church by given id
      operationId: getChurchBeliever
      parameters:
        - name: churchid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Believer Found
          schema:
            type: array
            items:
              $ref: "#/definitions/believer"
        "404":
          description: Church not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"

#Path to retrieve church Admins
  /church/{churchid}/admin:
    x-swagger-router-controller: church.controller
    get:
      tags: [Church]
      description: Return single church by given id
      operationId: getChurchAdmin
      parameters:
        - name: churchid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Admin Found
          schema:
            type: array
            items:
              $ref: "#/definitions/admin"
        "404":
          description: Admin not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# The End of Churches
#Path to retrieve church Believers  By age Range
  /church/{churchid}/{agefrom}/{ageto}/believers:
    x-swagger-router-controller: church.controller
    get:
      tags: [Church]
      description: Return single church by given id
      operationId: getChurchBelieverByAge
      parameters:
        - name: churchid
          required: true
          in: path
          type: string
        - name: agefrom
          required: true
          in: path
          type: number
        - name: ageto
          required: true
          in: path
          type: number
      responses:
        "200":
          description: Believer Found
          schema:
            type: array
            items:
              $ref: "#/definitions/believer"
        "404":
          description: Church not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# The End of Churches
# The Start of Info Marchavel
...
  /community:
    # binds a127 app logic to a route
    x-swagger-router-controller: community.controller
    get:
      tags: [Community]
      description: Returns list of communitys from Community Collection
      # used as the method name of the controller
      operationId: getAllCommunity
      responses:
        "200":
          description: Community Found
          schema:
            type: array
            items:
              $ref: "#/definitions/community"
        "404":
          description: Empty Collection
          schema:
            $ref: "#/definitions/message"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    post:
      tags: [Community]
      description: Creates new community entry
      consumes:
        - application/json
      produces:
        - application/json
      operationId: createCommunity
      parameters:
        - name: communitydata
          required: true
          in: body
          description: a new community details
          schema:
            $ref: "#/definitions/community"
      responses:
        "200":
          description: New community Created
          schema:
            $ref: "#/definitions/community"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Community]
      description: Remove all communityes in database
      operationId: removeAllCommunity
      responses:
        "200":
          description: Successfully removed all communityes
          schema:
            $ref: "#/definitions/message"
        default:
          description: Error
          schema:
            $ref: "#/definitions/message"
  /community/{communityid}:
    x-swagger-router-controller: community.controller
    get:
      tags: [Community]
      description: Return single community by given id
      operationId: getSingleCommunity
      parameters:
        - name: communityid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Community Found
          schema:
            $ref: "#/definitions/community"
        "404":
          description: Community not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Community]
      description: Remove single community by given id
      operationId: removeSingleCommunity
      parameters:
        - name: communityid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Successfully removed
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    patch:
      tags: [Community]
      description: Update by community id
      operationId: updateSingleCommunity
      parameters:
        - name: communityid
          in: path
          required: true
          type: string
        - name: community
          in: body
          description: The Updated community info
          required: true
          schema:
            $ref: "#/definitions/community"
      responses:
        "200":
          description: Community Found and Updated
          schema:
            $ref: "#/definitions/community"
        "404":
          description: Community not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
#Path to retrieve community's Believers
  /community/{communitycode}/believers:
    x-swagger-router-controller: community.controller
    get:
      tags: [Community]
      description: Return single community by given id
      operationId: getCommunityBeliever
      parameters:
        - name: communitycode
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Believer Found
          schema:
            type: array
            items:
              $ref: "#/definitions/believer"
        "404":
          description: Community not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
#Path to retrieve community's Believers By Age Range
  /community/{communitycode}/{agefrom}/{ageto}/believers:
    x-swagger-router-controller: community.controller
    get:
      tags: [Community]
      description: Return single community by given id
      operationId: getCommunityBelieverByAge
      parameters:
        - name: communitycode
          required: true
          in: path
          type: string
        - name: agefrom
          required: true
          in: path
          type: number
        - name: ageto
          required: true
          in: path
          type: number
      responses:
        "200":
          description: Believer Found
          schema:
            type: array
            items:
              $ref: "#/definitions/believer"
        "404":
          description: Community not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# The End of Community

# The Start of Contribution Marchavel
...
definitions:
...
  church:
    properties:
      _id:
        type: string
      churchcode:
        type: string
      churchname:
        type: string
      churchlocation:
        type: string
      diocode:
        type: string
      dioname:
        type: string
      churchleader:
        type: object
        properties:
          leadername:
            type: string
          position:
            type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime
# Community Schema Definition
  community:
    properties:
      _id:
        type: string
      communitycode:
        type: string
      communityname:
        type: string
      communitylocation:
        type: string
      churchid:
        type: string
      churchcode:
        type: string
      churchname:
        type: string
      diocode:
        type: string
      dioname:
        type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime

# Contribution Schema Definition
  contribution:
    properties:
      _id:
        type: string
      contributionname:
        type: string
      contributionpurpose:
        type: string
      startin:
        type: string
      deadline:
        type: string
      amount:
        type: number
      level:
        type: string
      affectedgroup:
        type: array
        items:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            code:
              type: string
            location:
              type: string
      believerinclude:
        type: string
      agelimit:
        type: object
        properties:
          low:
            type: string
          high:
            type: string
      churchid:
        type: string
      churchcode:
        type: string
      churchname:
        type: string
      diocode:
        type: string
      dioname:
        type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime

# Contributor Schema Definition
...
  believer:
    properties:
      _id:
        type: string
      believercode:
        type: string
      title:
        type: string
      firstname:
        type: string
      middlename:
        type: string
      lastname:
        type: string
      phonenumber:
        type: string
      device_id:
        type: string
      diocode:
        type: string
      dioname:
        type: string
      churchid:
        type: string
      churchcode:
        type: string
      churchname:
        type: string
      communityid:
        type: string
      communitycode:
        type: string
      communityname:
        type: string
      profilepicture:
        type: string
      email:
        type: string
      gender:
        type: string
      dob:
        type: object
        properties:
          year:
            type: number
          month:
            type: number
          day:
            type: number
      status:
        type: string
      healthinfo:
        type: string
      nationality:
        type: string
      jobtitle:
        type: string
      regdate:
        type: string
        format: datetime
      baptismstatus:
        type: string
      maritalstatus:
        type: string
      kipaimarastatus:
        type: string
      participationlevel:
        type: array
        items:
          type: string
      residenceadress:
        type: object
        properties:
          street:
            type: string
          road:
            type: string
          district:
            type: string
          region:
            type: string
      relatives:
        type: array
        items:
          type: object
          properties:
            relation:
              type: string
            person:
              type: string
      hobbies:
        type: array
        items:
          type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime
#Relative model Definition
...
  admin:
    properties:
      _id:
        type: string
      admincode:
        type: string
      title:
        type: string
      firstname:
        type: string
      middlename:
        type: string
      lastname:
        type: string
      phonenumber:
        type: string
      device_id:
        type: string
      diocode:
        type: string
      dioname:
        type: string
      churchid:
        type: string
      churchcode:
        type: string
      churchname:
        type: string
      communitycode:
        type: string
      communityname:
        type: string
      profilepicture:
        type: string
      email:
        type: string
      gender:
        type: string
      dob:
        type: object
        properties:
          year:
            type: number
          month:
            type: number
          day:
            type: number
      status:
        type: string
      healthinfo:
        type: string
      nationality:
        type: string
      jobtitle:
        type: string
      regdate:
        type: string
        format: datetime
      baptismstatus:
        type: string
      maritalstatus:
        type: string
      participationlevel:
        type: array
        items:
          type: string
      residenceadress:
        type: object
        properties:
          street:
            type: string
          road:
            type: string
          district:
            type: string
          region:
            type: string
      relatives:
        type: array
        items:
          type: object
          properties:
            relation:
              type: string
            person:
              type: string
      hobbies:
        type: array
        items:
          type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime
#Login Model Defition
...
  message:
    required:
      - success
      - message
    properties:
      success:
        type: boolean
      message:
        type: string
# Info Model Definition
...